+++ sapl/sessao/forms.py
@@ -23,7 +23,7 @@ from sapl.utils import (RANGE_DIAS_MES, RANGE_MESES,
                         autor_modal, timezone, choice_anos_com_sessaoplenaria,
                         FileFieldCheckMixin)
 
-from .models import (Bancada, Bloco, ExpedienteMateria, JustificativaAusencia,
+from .models import (Bancada, ExpedienteMateria, JustificativaAusencia,
                      Orador, OradorExpediente, OrdemDia, PresencaOrdemDia, SessaoPlenaria,
                      SessaoPlenariaPresenca, TipoResultadoVotacao,
                      OcorrenciaSessao, RetiradaPauta, TipoRetiradaPauta, OradorOrdemDia, ORDENACAO_RESUMO,
@@ -314,41 +314,6 @@ class BancadaForm(ModelForm):
             nome=bancada.nome
         )
         return bancada
-
-
-class BlocoForm(ModelForm):
-
-    class Meta:
-        model = Bloco
-        fields = ['nome', 'partidos', 'data_criacao',
-                  'data_extincao', 'descricao']
-
-    def clean(self):
-        super(BlocoForm, self).clean()
-
-        if not self.is_valid():
-            return self.cleaned_data
-
-        if self.cleaned_data['data_extincao']:
-            if (self.cleaned_data['data_extincao'] <
-                    self.cleaned_data['data_criacao']):
-                msg = _('Data de extinção não pode ser menor que a de criação')
-                raise ValidationError(msg)
-        return self.cleaned_data
-
-    @transaction.atomic
-    def save(self, commit=True):
-        bloco = super(BlocoForm, self).save(commit)
-        content_type = ContentType.objects.get_for_model(Bloco)
-        object_id = bloco.pk
-        tipo = TipoAutor.objects.get(content_type=content_type)
-        Autor.objects.create(
-            content_type=content_type,
-            object_id=object_id,
-            tipo=tipo,
-            nome=bloco.nome
-        )
-        return bloco
 
 
 class ExpedienteMateriaForm(ModelForm):
